+++ puzzles/day4/Day4.h
@@ -9,8 +9,8 @@
 
 class Day4 final : public Solution {
 public:
-    std::pair<std::string, int> solutionA() override;
-    std::pair<std::string, int> solutionB() override;
+    std::pair<std::string, long long> solutionA() override;
+    std::pair<std::string, long long> solutionB() override;
 
 private:
     std::string inputALocation = "../puzzles/day4/input.txt";
